pyproject.toml
[project]
name = "local-media-server"
version = "1.0.0"
description = "本地媒体服务器 - 基于 Python + FastAPI 的轻量级文件服务系统"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Local Media Server Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "python-magic>=0.4.27",
    "Pillow>=10.1.0",
    "rarfile>=4.1.0",
    "py7zr>=0.20.0",
    "ffmpeg-python>=0.2.0",
    "cachetools>=5.3.0",
    "diskcache>=5.6.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "natsort>=8.4.0",
    "PyYAML>=6.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "hypothesis>=6.92.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=backend --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
